Updated Travis build
This change prevents out-of-memory errors by taking gunicorn out of the equation. Each part of the script starts a fresh container that only runs the script command, instead of running everything alongside a container that also runs gunicorn.

Additionally, `make test` assumes it is running in an environment with the appropriate environment variables configured.